split aql-modify-noncluster testsuite into more digesteable parts
